Move `subsidies-decimal-point` to the utils folder
It's only used as a util in the code, not a template helper.
